feat(alert): configurable original message deletion

+++ modules/payments-bridge/sear-payments-bridge-firco/src/main/java/com/silenteight/payments/bridge/firco/core/alertmessage/service/AlertMessageStatusService.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import com.silenteight.payments.bridge.firco.core.alertmessage.model.AlertMessageStatus;
 
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Service;
 
 import java.time.Clock;
@@ -13,6 +14,7 @@ import java.util.UUID;
 import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 
+@EnableConfigurationProperties(AlertMessageProperties.class)
 @RequiredArgsConstructor
 @Service
 @Slf4j
@@ -20,6 +22,7 @@ class AlertMessageStatusService {
 
   private final AlertMessageStatusRepository repository;
   private final AlertMessagePayloadRepository payloadRepository;
+  private final AlertMessageProperties alertMessageProperties;
 
   @Setter
   private Clock clock = Clock.systemUTC();
@@ -43,7 +46,8 @@ class AlertMessageStatusService {
 
     repository.save(entity);
 
-    if (destinationStatus.isFinal()) {
+    if (destinationStatus.isFinal() &&
+        alertMessageProperties.isOriginalMessageDeletedAfterRecommendation()) {
       payloadRepository.deleteByAlertMessageId(alertMessageId);
     }
   }
